extract CSL rendering to standalone object
Allows us to more easily test exact rendered html, among other things.
